LNIT-32-스터디-그룹-찜-토글-API-구현 (#21)
* feat: Dibs 기능 및 관련 DTO, 서비스, 컨트롤러 구현

- DibsController 및 DibsService를 추가하여 찜하기 기능을 구현
- DibsDto를 생성하여 찜하기 응답 형식을 정의
- StudyGroupController 및 관련 서비스에서 ID 타입을 Integer에서 Long으로 변경하여 일관성 유지
- DibsRepository 및 관련 테스트 케이스 추가
- 기존 코드에서 Integer ID를 Long으로 변경하여 데이터베이스와의 호환성 향상
- DibsService의 toggleDibs 메서드에서 찜하기 추가 및 삭제 로직 구현
- 관련 테스트 케이스를 통해 기능 검증 및 안정성 확보

* fix: 테스트 코드에서 사용자 ID 타입을 Long으로 변경

- DibsServiceTest, StudyGroupApplicationServiceTest, StudyGroupServiceTest, UserServiceTest에서 사용자 ID를 Long 타입으로 수정
- 관련된 모든 테스트 케이스에서 ID를 Long으로 일관되게 사용하도록 업데이트

* feat: 스터디 그룹 찜 토글 API에 Swagger 문서화 추가

- DibsController에 Swagger @Tag 및 @Operation 어노테이션 추가
- API 설명 및 요약을 통해 문서화 개선

+++ src/main/java/com/depth/learningcrew/domain/studygroup/controller/StudyGroupController.java
@@ -42,7 +42,7 @@ public class StudyGroupController {
   @PatchMapping(value = "/{groupId}")
   @Operation(summary = "스터디 그룹 정보 수정", description = "owner만 스터디 그룹의 정보를 수정할 수 있습니다. 모든 정보를 하나의 요청으로 전송합니다.")
   public StudyGroupDto.StudyGroupResponse updateStudyGroup(
-      @PathVariable Integer groupId,
+      @PathVariable Long groupId,
       @Valid @ModelAttribute @ParameterObject StudyGroupDto.StudyGroupUpdateRequest request,
       @AuthenticationPrincipal UserDetails userDetails) {
     return studyGroupService.updateStudyGroup(groupId, request, userDetails);
@@ -52,7 +52,7 @@ public class StudyGroupController {
     @Operation(summary = "스터디 그룹 상세 조회", description = "스터디 그룹 ID를 통해 스터디 그룹의 상세 정보를 조회합니다.")
     @ApiResponse(responseCode = "200", description = "스터디 그룹 조회 성공")
     public StudyGroupDto.StudyGroupDetailResponse getStudyGroup(
-            @PathVariable Integer id,
+            @PathVariable Long id,
             @AuthenticationPrincipal UserDetails userDetails){
         return studyGroupService.getStudyGroupDetail(id, userDetails);
       }
