+++ src/main/java/com/depth/learningcrew/domain/studygroup/service/StudyGroupApplicationService.java
@@ -34,7 +34,7 @@ public class StudyGroupApplicationService {
 
   @Transactional(readOnly = true)
   public PagedModel<ApplicationDto.ApplicationResponse> getApplicationsByGroupId(
-      Integer groupId,
+      Long groupId,
       ApplicationDto.SearchConditions searchConditions,
       UserDetails userDetails,
       Pageable pageable) {
@@ -54,7 +54,7 @@ public class StudyGroupApplicationService {
   }
 
   @Transactional
-  public ApplicationDto.ApplicationResponse joinStudyGroup(Integer groupId, UserDetails userDetails) {
+  public ApplicationDto.ApplicationResponse joinStudyGroup(Long groupId, UserDetails userDetails) {
     StudyGroup studyGroup = studyGroupRepository.findById(groupId)
         .orElseThrow(() -> new RestException(ErrorCode.GLOBAL_NOT_FOUND));
 
@@ -72,7 +72,7 @@ public class StudyGroupApplicationService {
   }
 
   @Transactional
-  public ApplicationDto.ApplicationResponse approveApplication(Integer groupId, Integer userId,
+  public ApplicationDto.ApplicationResponse approveApplication(Long groupId, Long userId,
       UserDetails ownerDetails) {
     StudyGroup studyGroup = studyGroupRepository.findById(groupId)
         .orElseThrow(() -> new RestException(ErrorCode.GLOBAL_NOT_FOUND));
@@ -103,8 +103,7 @@ public class StudyGroupApplicationService {
   }
 
   @Transactional
-  public ApplicationDto.ApplicationResponse rejectApplication(Integer groupId, Integer userId,
-      UserDetails ownerDetails) {
+  public ApplicationDto.ApplicationResponse rejectApplication(Long groupId, Long userId, UserDetails ownerDetails) {
     if (!studyGroupRepository.existsById(groupId)) {
       throw new RestException(ErrorCode.GLOBAL_NOT_FOUND);
     }
